Move the iu meta data preprocessing into seperate step
